cmake will only run properly named c10 tests
Pull Request resolved: https://github.com/pytorch/pytorch/pull/98710

For the purpose of our Bazel and Meta-internal macros tests, we want
to create a single binary that can verify the different
configurations. CMake would see this file and try to run it and fail
in Windows which uses different values.

But we don't care about verifying this in CMake since it's not part of
the build unification effort.

In order to do this, we have to rename the SmallVectorTest to match
the naming convention of the rest of the c10 tests.
ghstack-source-id: 185576029

Differential Revision: [D44823440](https://our.internmc.facebook.com/intern/diff/D44823440/)